refactor: Convert Javascript files into Typescript (#845)
* cheaper debug build

* index.ts, App.tsx

* TS for navigators

* TS NovelScreen

* TS Reader Screen

* TS More Screen

* TS Source Screen

* TS Browse Settings Screen

* TS Global Search Screen

* fix: useNavigation for more stack

* fix: Composite Types for More Screen

* TS Browse + Migrate Screen

* TS Settings Screen

* TS util

* TS services

* fix: Library query

* TS Components

* refactor: settingsSliceV1

* refactor: Preference reducer

* TS for MAL

* refactor: Native modules

* Not TS the entry file

* fix: zip folder entry, special file name

* fix: plugin is undefined if 'local'

* fix: add param for morestack navigation

* fix: Component is bool

* fix: novel reducer

* fix: not updating local novels

+++ android/app/src/main/java/com/rajarsheechatterjee/EpubParser/EpubParser.java
@@ -138,6 +138,10 @@ public class EpubParser extends ReactContextBaseJavaModule {
         return doc;
     }
 
+    public String escapeFilePath(String filePath){
+        return filePath.replaceAll(":", "\uA789");
+    }
+
     public String unzip(String epubFilePath, String epubDirPath) throws Exception {
         ZipInputStream zis = new ZipInputStream(new FileInputStream(epubFilePath));
         ZipEntry zipEntry;
@@ -145,15 +149,23 @@ public class EpubParser extends ReactContextBaseJavaModule {
         String novelCover = null;
         byte[] buffer = new byte[4096];
         while ((zipEntry = zis.getNextEntry()) != null) {
-            if(Pattern.compile(".*cover.*\\.(png|jpeg|jpg)", Pattern.CASE_INSENSITIVE).matcher(zipEntry.getName()).matches()){
-                novelCover = zipEntry.getName();
+            String escapedFilePath = escapeFilePath(zipEntry.getName());
+            if(Pattern.compile(".*cover.*\\.(png|jpeg|jpg)", Pattern.CASE_INSENSITIVE).matcher(escapedFilePath).matches()){
+                novelCover = escapedFilePath;
             }
-            File newFile = new File(epubDirPath, zipEntry.getName());
+            File newFile = new File(epubDirPath, escapedFilePath);
+            if(newFile.exists()) continue;
             newFile.getParentFile().mkdirs();
             FileOutputStream fos = new FileOutputStream(newFile);
-            while ((len = zis.read(buffer)) > 0)
+            int writeCount = 0;
+            while ((len = zis.read(buffer)) > 0) {
+                writeCount += 1;
                 fos.write(buffer, 0, len);
+            }
             fos.close();
+            if(writeCount == 0){
+                newFile.delete();
+            }
         }
         zis.closeEntry();
         zis.close();
